Move find_peaks from gammapy.image to gammapy.detect
